refactoring to remove globals and implement types
